Abstracted poll handling to a PollHandler object, removed DB query on single vote check favoring an in-memory lookup
